backend/requirements.txt
fastapi==0.68.1
python-multipart==0.0.5
pytesseract==0.3.8
pdf2image==1.16.0
PyPDF2==3.0.1
Pillow==8.3.2
uvicorn==0.15.0
python-magic==0.4.24
openai

